Routing clean up
