+++ components/locked-tab.tsx
@@ -1,0 +1,3 @@
+const LockedTab = () => null;
+
+export { LockedTab };
